requirements.txt
Flask~=2.2
Flask_SQLAlchemy~=3.0
Flask_Limiter~=3.3
Mastodon.py~=1.8
